% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS100_do_it.R
\name{do_it}
\alias{do_it}
\title{Repeats a process many times and returns a vector of results}
\usage{
do_it(n = 1L)
}
\arguments{
\item{n}{The number of times to repeat the process.}

\item{null_dist}{A vector of data consistent with a null hypothesis.}

\item{lower.tail}{A Boolean indicating whether one should return the
proportion of points less than or equal to the obs_stat value.}
}
\description{
Repeats a process many times and returns a vector of results
}
\note{
This is a simplified version of the mosiac do() that returns a vector only.
The idea and code are based on the mosaic do() function, which can be found
on github page at: https://github.com/ProjectMOSAIC/mosaic/blob/master/R/do.R
The code uses S4 objects.
}
\examples{
 many_hellos <- do_it(10) * { "hello" }

}
