% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS100_get_statistics.R
\name{get_chisqr_stat}
\alias{get_chisqr_stat}
\title{Calculates the chi-square statistic}
\usage{
get_chisqr_stat(observed_counts, expected_proportions)
}
\arguments{
\item{observed_counts}{A vector of observed count data.}

\item{expected_proportions}{A vector indicating the expected probability that
data comes from a given group.}
}
\description{
Calculates the chi-square statistic
}
\examples{
 set.seed(100)
 observed_counts <- c(138, 99, 106, 115, 104, 164)
 expected_proportions = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6)
 get_chisqr_stat(observed_counts, expected_proportions)

}
