% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS100_get_statistics.R
\name{get_F_stat}
\alias{get_F_stat}
\title{Calculates the F statistic}
\usage{
get_F_stat(data, grouping)
}
\arguments{
\item{data}{A vector of quantitative data.}

\item{grouping}{A vector of categorical data indicating which group each
value of the quantitative data belongs to.}
}
\description{
Calculates the F statistic
}
\examples{
 set.seed(100)
 the_data <- c(rnorm(10, -3), rnorm(10, 0), rnorm(10, 3))
 grouping <- c(rep("A", 10), rep("B", 10), rep("C", 10))
 get_F_stat(the_data, grouping)

}
