% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS100_get_MAD_F_chisq_stat.R
\name{get_MAD_stat}
\alias{get_MAD_stat}
\title{Calculates the MAD statistic}
\usage{
get_MAD_stat(data, grouping)
}
\arguments{
\item{data}{A vector of quantitative data.}

\item{grouping}{A vector of categorical data indicating which group each
value of the quantitative data belongs to.}
}
\description{
The MAD statistic is the Mean Absolute Difference between the means of
several groups.
}
\examples{
 set.seed(100)
 the_data <- c(rnorm(10, -3), rnorm(10, 0), rnorm(10, 3))
 grouping <- c(rep("A", 10), rep("B", 10), rep("C", 10))
 get_MAD_stat(the_data, grouping)

}
