% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS100_get_statistics.R
\name{get_proportion}
\alias{get_proportion}
\title{Calculates proportion of values in a particular category}
\usage{
get_proportion(v, category_name)
}
\arguments{
\item{v}{A vector of categorical data.}

\item{category_name}{A string specifying the name of a category. The
proportion of values in that category will be returned.}
}
\description{
Calculates proportion of values in a particular category
}
\examples{
 set.seed(100)
 sprinkle_sample <- get_sprinkle_sample(100)
 get_proportion(sprinkle_sample, "red")

}
