% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS100_generate_simulated_samples.R
\name{get_sprinkle_sample}
\alias{get_sprinkle_sample}
\title{Generate a random sample of sprinkle colors}
\usage{
get_sprinkle_sample(n)
}
\arguments{
\item{n}{The sample size.}
}
\description{
This function generates a fictional sample of data of sprinkles.
}
\examples{
 # Generate a sample from 10 sprinkle colors
 get_sprinkle_sample(10)

}
