% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS100_generate_simulated_samples.R
\name{rflip_sequence}
\alias{rflip_sequence}
\title{Generates a sequence of coin flips}
\usage{
rflip_sequence(num_flips = 1, prob = 0.5, output_type = "name")
}
\arguments{
\item{num_flips}{The number of times to flip the coin.}

\item{prob}{The probability of generated a "heads" on each flip.}

\item{output_type}{A string that must either be set to "numeric", "name" or
"long name". If this string is set to "numeric" the output values are 0's
and 1's, where a 1 is a "heads". If this string is set to "name" the output
values are "H" and "T", where a "H" is a "heads". If this string is set
to "long name" the output values are "Heads" and "Tails".}
}
\description{
Generates a sequence of coin flips
}
\examples{
 set.seed(100)
 rflip_sequence(10, output_type = "numeric")

}
